Add latest teleporter entries. No longer adding whitelist and blacklist separately
I will no longer be adding the white and black lists separately since the pihole UI makes it difficult to extract only the domains from the page.

The domains are all stored in the teleporter files anyway.

I will only be adding the blocklists I use going forward as that is a lot easier to do and still helpful